--- beginner/utilities/draw-modem-new/.cargo/config
@@ -1,8 +1,0 @@
-[build]
-target = "thumbv7em-none-eabihf"
-
-[target.thumbv7em-none-eabihf]
-runner = "arm-none-eabi-gdb -tui -q -x debug.gdb"
-rustflags = [
-    "-C", "link-arg=-Tlink.x",
-]
