--- beginner/utilities/draw-modem-new/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "better-logging"
-version = "0.1.0"
-authors = ["James Munns <james.munns@ferrous-systems.com>"]
-edition = "2018"
-
-[dependencies]
-panic-ramdump    = "0.1.0"
-nb               = "0.1.1"
-cortex-m-rtfm    = "0.4.1"
-nrf52832-pac     = "0.6.0"
-dw1000           = "0.1.0"
-embedded-hal     = "0.2.2"
-heapless         = "0.4.2"
-postcard = "0.3.2"
-
-[dependencies.nrf52-bin-logger]
-version = "0.3.0"
-features = ["unstable"]
-
-[dependencies.ssmarshal]
-version       = "=1.0.0"
-default-features = false
-
-[dependencies.dwm1001]
-version = "0.1.1"
-features = [ "dev", "rt" ]
-
-# Local workspace deps
-
-[dependencies.protocol]
-path = "../../protocol"
-
-[profile.release]
-opt-level = 3
-debug = true
-lto = true
-incremental = false
-codegen-units = 1
-
-[patch.crates-io.dw1000]
-git = "https://github.com/braun-robotics/rust-dw1000"
-rev = "036dc67b3147417c49318b840f4d769f8045967e"
