--- beginner/utilities/draw-modem-new/debug.gdb
@@ -1,6 +1,0 @@
-target remote :2331
-set backtrace limit 32
-load
-monitor reset
-break main
-continue
