--- beginner/utilities/draw-server-new/.cargo/config
@@ -1,3 +1,0 @@
-[build]
-#target = "x86_64-unknown-linux-gnu"
-target = "x86_64-apple-darwin"
