--- beginner/utilities/draw-server-new/draw-freedraw.ron
@@ -1,20 +1,0 @@
-(
-    serial: (
-        timeout_ms: 100,
-        baudrate: 115_200,
-        port: "/dev/tty.usbmodem0007600866611",
-    ),
-    squares: (
-        host: "http://localhost",
-        port: 8000,
-    ),
-    board: (
-        total_board: (
-            x: ( start: 1, end: 32 ),
-            y: ( start: 1, end: 32 ),
-        ),
-        mode: FreeDraw(
-            clear_interval: ( secs: 60, nanos: 0 ),
-        ),
-    )
-)
