--- beginner/utilities/draw-server-new/draw-roundrobin.ron
@@ -1,22 +1,0 @@
-(
-    serial: (
-        timeout_ms: 100,
-        baudrate: 115_200,
-        port: "/dev/ttyACM0",
-    ),
-    squares: (
-        host: "http://localhost",
-        port: 8000,
-    ),
-    board: (
-        total_board: (
-            x: ( start: 1, end: 32 ),
-            y: ( start: 1, end: 32 ),
-        ),
-        mode: RoundRobin (
-            turn_interval: ( secs: 5, nanos: 0 ),
-            notify_interval: ( secs: 0, nanos: 500_000_000 ),
-            players: [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16 ],
-        ),
-    )
-)
